Add double click event for tray icon
The `mainWindow` was not opening when double clicking on the tray icon
so I added a new event to the `appIcon` tray object. Now it will show
the `mainWindow` whenever the user double clicks on the tray icon.

The code had a few indentation problems so I decided to format it in a
way that makes it more readable.

With the new option to hide to tray, the window space became too tight
to fit the options under the Reset and Start button. So I tweaked the
`mainWindow`, `width` and `height` properties.

Co-authored-by: cisnejo <cisnerosjoseph122@gmail.com>