Replace `span` to `label` for the checkbox input
With the `<span>` tag it was not possible to click on the text to
select the checkbox for hiding the window to tray. So I changed to the
`<label>` tag with the `for` attribute linking it to the checkbox input.
Now whenever the user clicks on the checkbox or the label it will check
or uncheck it.

Due to the new with and height of the window, I added a `mt-3` class to
add a margin on top of the div container to space out from the buttons.

Co-authored-by: cisnejo <cisnerosjoseph122@gmail.com>