Fix the tray icon not showing on production build
After packaging the application, the path of the tray icon would no
longer be the same as in the source files. To fix this, it was
implemented a ternary statement to check if the app is packaged or not,
if it is packaged then it will look for the icon in
`process.ResourcesPath`.

Co-authored-by: cisnejo <cisnerosjoseph122@gmail.com>